requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124

numpy
pandas
matplotlib
pillow
scikit-learn
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm
